memvidBeta/chat_app/requirements.txt
python-telegram-bot>=13.7
python-dotenv>=0.19.2
requests>=2.28.1
memvid
SQLAlchemy>=1.4.46
aiosqlite
pydantic>=1.9.0
